requirements.txt
streamlit
pymupdf
pytesseract
pillow
scikit-learn
groq
python-dotenv
